requirements.txt
django-model-utils>=2.0
django-two-factor-auth
social-auth-app-django
